# apps/orion-chat/shared/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "orion-agent",
    "description": "Multi-LLM agent harness FastAPI layer. WebSocket /chat/stream + REST /sessions + JWT /auth/login.",
    "version": "0.1.0"
  },
  "paths": {
...
    "/roles": {
      "get": {
        "summary": "List Roles",
        "operationId": "list_roles_roles_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/RoleSummary"
                  },
                  "type": "array",
                  "title": "Response List Roles Roles Get"
                }
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      }
    },
    "/roles/{name}": {
      "get": {
        "summary": "Get Role",
        "operationId": "get_role_roles__name__get",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "name",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Name"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RoleDetail"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "put": {
        "summary": "Put Role",
        "operationId": "put_role_roles__name__put",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "name",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Name"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RoleWriteBody"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RoleDetail"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "delete": {
        "summary": "Delete Role",
        "operationId": "delete_role_roles__name__delete",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "name",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Name"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "additionalProperties": {
                    "type": "boolean"
                  },
                  "title": "Response Delete Role Roles  Name  Delete"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/schedules": {
      "get": {
        "summary": "List Schedules",
        "operationId": "list_schedules_schedules_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/ScheduleSummary"
                  },
                  "type": "array",
                  "title": "Response List Schedules Schedules Get"
                }
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      },
      "post": {
        "summary": "Create Schedule",
        "operationId": "create_schedule_schedules_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ScheduleBody"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ScheduleSummary"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      }
    },
    "/schedules/{sched_id}": {
      "put": {
        "summary": "Update Schedule",
        "operationId": "update_schedule_schedules__sched_id__put",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "sched_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Sched Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ScheduleBody"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ScheduleSummary"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "delete": {
        "summary": "Delete Schedule",
        "operationId": "delete_schedule_schedules__sched_id__delete",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "sched_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Sched Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "additionalProperties": {
                    "type": "boolean"
                  },
                  "title": "Response Delete Schedule Schedules  Sched Id  Delete"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "RoleDetail": {
        "properties": {
          "name": {
            "type": "string",
            "title": "Name"
          },
          "description": {
            "type": "string",
            "title": "Description"
          },
          "default_disabled_tools": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Default Disabled Tools"
          },
          "default_permission_mode": {
            "anyOf": [
              {
                "type": "string",
                "enum": [
                  "ask",
                  "act"
                ]
              },
              {
                "type": "null"
              }
            ],
            "title": "Default Permission Mode"
          },
          "editable": {
            "type": "boolean",
            "title": "Editable"
          },
          "body": {
            "type": "string",
            "title": "Body"
          }
        },
        "type": "object",
        "required": [
          "name",
          "description",
          "editable",
          "body"
        ],
        "title": "RoleDetail"
      },
      "RoleSummary": {
        "properties": {
          "name": {
            "type": "string",
            "title": "Name"
          },
          "description": {
            "type": "string",
            "title": "Description"
          },
          "default_disabled_tools": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Default Disabled Tools"
          },
          "default_permission_mode": {
            "anyOf": [
              {
                "type": "string",
                "enum": [
                  "ask",
                  "act"
                ]
              },
              {
                "type": "null"
              }
            ],
            "title": "Default Permission Mode"
          },
          "editable": {
            "type": "boolean",
            "title": "Editable"
          }
        },
        "type": "object",
        "required": [
          "name",
          "description",
          "editable"
        ],
        "title": "RoleSummary"
      },
      "RoleWriteBody": {
        "properties": {
          "description": {
            "type": "string",
            "maxLength": 2000,
            "title": "Description",
            "default": ""
          },
          "body": {
            "type": "string",
            "maxLength": 200000,
            "title": "Body",
            "default": ""
          },
          "default_disabled_tools": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Default Disabled Tools"
          },
          "default_permission_mode": {
            "anyOf": [
              {
                "type": "string",
                "enum": [
                  "ask",
                  "act"
                ]
              },
              {
                "type": "null"
              }
            ],
            "title": "Default Permission Mode"
          }
        },
        "type": "object",
        "title": "RoleWriteBody"
      },
      "ScheduleBody": {
        "properties": {
          "name": {
            "type": "string",
            "maxLength": 200,
            "minLength": 1,
            "title": "Name"
          },
          "cron_expr": {
            "type": "string",
            "maxLength": 120,
            "minLength": 1,
            "title": "Cron Expr"
          },
          "trigger_type": {
            "type": "string",
            "enum": [
              "prompt",
              "skill"
            ],
            "title": "Trigger Type",
            "default": "prompt"
          },
          "payload": {
            "type": "string",
            "title": "Payload",
            "default": ""
          },
          "enabled": {
            "type": "boolean",
            "title": "Enabled",
            "default": true
          },
          "target_session_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Target Session Id"
          }
        },
        "type": "object",
        "required": [
          "name",
          "cron_expr"
        ],
        "title": "ScheduleBody"
      },
      "ScheduleSummary": {
        "properties": {
          "id": {
            "type": "string",
            "title": "Id"
          },
          "name": {
            "type": "string",
            "title": "Name"
          },
          "cron_expr": {
            "type": "string",
            "title": "Cron Expr"
          },
          "trigger_type": {
            "type": "string",
            "title": "Trigger Type"
          },
          "payload": {
            "type": "string",
            "title": "Payload"
          },
          "enabled": {
            "type": "boolean",
            "title": "Enabled"
          },
          "target_session_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Target Session Id"
          },
          "next_run_at": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Next Run At"
          }
        },
        "type": "object",
        "required": [
          "id",
          "name",
          "cron_expr",
          "trigger_type",
          "payload",
          "enabled"
        ],
        "title": "ScheduleSummary"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      },
...
    },
    "securitySchemes": {
      "HTTPBearer": {
        "type": "http",
        "scheme": "bearer"
      }
    }
  }
}